Add saturate step

+++ src/pipeline/steps/removeBackground.ts
@@ -2,9 +2,12 @@
 import removeBackground from "@imgly/background-removal-node"
 import path from 'path';
 import fs from 'fs'
+import { fileHash } from "../util/FileHash";
 
 const relativePath = 'public/out'
 const outDir = path.join(process.cwd(), relativePath);
+
+const cache = new Map<string, string>()
 
 const config: any = {
     progress: (key: string, current: string, total: string) => {
@@ -14,13 +17,21 @@ const config: any = {
 
 export const removeBg = async (imagePath: string, rawImagePath: string): Promise<string> => {
     const inPath = path.join(process.cwd(), imagePath)
+    const imageHash = await fileHash(inPath);
+
+    if (cache.has(imageHash) && fs.existsSync(cache.get(imageHash)!!)) {
+      console.log('We have already removed the background for this image! Using cached result...')
+      return cache.get(imageHash)!!
+    }
+
     const relativeFilename = `bgimg${Date.now()}.png`
     const outPath = `${outDir}/${relativeFilename}`
-    
-    console.log('removing bg')
     const blob: Blob = await removeBackground(inPath, config)
     const buffer = await blob.arrayBuffer();
-    console.log('finished removing bg')
+    const result = `${relativePath}/${relativeFilename}`
     await fs.promises.writeFile(outPath, Buffer.from(buffer));
-    return `${relativePath}/${relativeFilename}`;
+
+    cache.set(imageHash, result)
+
+    return result;
 }
